resolve merge conflits

+++ tsfel/feature_extraction/calc_features.py
@@ -168,7 +168,7 @@ def calc_features(wind_sig, dict_features, fs, **kwargs):
         Input from which features are computed, window
     dict_features : dict
         Dictionary with features
-    fs : int or None
+    fs : float or None
         Sampling frequency
     \**kwargs:
         * *features_path* (``string``) --
